move and finalize
