+++ crates/papyrus_gateway/src/test_utils.rs
@@ -4,7 +4,8 @@ use papyrus_storage::test_utils::get_test_storage;
 use papyrus_storage::StorageWriter;
 use starknet_api::core::ChainId;
 
-use crate::{GatewayConfig, JsonRpcServerV0_3_0Impl, JsonRpcV0_3_0Server};
+use crate::api::JsonRpcServerImpl;
+use crate::GatewayConfig;
 
 pub fn get_test_gateway_config() -> GatewayConfig {
     GatewayConfig {
@@ -15,18 +16,18 @@ pub fn get_test_gateway_config() -> GatewayConfig {
     }
 }
 
-pub(crate) fn get_test_rpc_server_and_storage_writer()
--> (RpcModule<JsonRpcServerV0_3_0Impl>, StorageWriter) {
+pub(crate) fn get_test_rpc_server_and_storage_writer<T: JsonRpcServerImpl>()
+-> (RpcModule<T>, StorageWriter) {
     let (storage_reader, storage_writer) = get_test_storage();
     let config = get_test_gateway_config();
     (
-        JsonRpcServerV0_3_0Impl {
-            chain_id: config.chain_id,
+        T::new(
+            config.chain_id,
             storage_reader,
-            max_events_chunk_size: config.max_events_chunk_size,
-            max_events_keys: config.max_events_keys,
-        }
-        .into_rpc(),
+            config.max_events_chunk_size,
+            config.max_events_keys,
+        )
+        .into(),
         storage_writer,
     )
 }
