--- crates/papyrus_gateway/src/version_config/mod.rs
@@ -1,4 +1,0 @@
-#[allow(clippy::module_inception)]
-pub mod version_config;
-#[cfg(test)]
-mod version_config_test;
