+++ crates/papyrus_monitoring_gateway/Cargo.toml
@@ -18,4 +18,4 @@ url.workspace = true
 
 [dev-dependencies]
 papyrus_storage = { path = "../papyrus_storage", features = ["testing"] }
-tower = { version = "0.4", features = ["util"] }
+tower = { workspace = true, features = ["util"] }
